+++ src/NativeUsbLib/UsbDevice.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Runtime.InteropServices;
+using NativeUsbLib.WinApis;
 
 namespace NativeUsbLib
 {
